+++ Whistle/UI/Screens/Profile/MyProfile/MyProfileView.swift
@@ -1,8 +1,8 @@
 //
-//  SEMyProfileView.swift
+//  MyProfileView.swift
 //  Whistle
 //
-//  Created by ChoiYujin on 10/10/23.
+//  Created by ChoiYujin on 8/29/23.
 //
 
 import AVKit
@@ -12,24 +12,31 @@ import GoogleSignInSwift
 import Kingfisher
 import SwiftUI
 
-// MARK: - SEMyProfileView
+// MARK: - profileTabCase
 
-struct SEMyProfileView: View {
+public enum profileTabCase: String {
+  case myVideo
+  case bookmark
+}
+
+// MARK: - MyProfileView
+
+struct MyProfileView: View {
   @StateObject var userAuth = UserAuth.shared
-  @StateObject private var tabbarModel = TabbarModel.shared
   @StateObject var apiViewModel = APIViewModel.shared
+  @StateObject private var tabbarModel = TabbarModel.shared
   @StateObject private var toastViewModel = ToastViewModel.shared
   @StateObject var alertViewModel = AlertViewModel.shared
 
   @State var isShowingBottomSheet = false
   @State var tabbarDirection: CGFloat = -1.0
   @State var tabSelection: profileTabCase = .myVideo
-
   @State var bottomSheetPosition: BottomSheetPosition = .hidden
   @State var offsetY: CGFloat = 0
-  @Binding var isFirstProfileLoaded: Bool
 
+  @Binding var isFirstProfileLoaded: Bool
   let processor = BlurImageProcessor(blurRadius: 10)
+
   var body: some View {
     ZStack {
       if bottomSheetPosition == .absolute(420) {
@@ -61,8 +68,8 @@ struct SEMyProfileView: View {
           glassProfile(
             cornerRadius: profileCornerRadius,
             overlayed: profileInfo())
-            .frame(height: 278 + (146 * progress))
-            .padding(.bottom, 8)
+            .frame(height: 418 + (240 * progress))
+            .padding(.bottom, 12)
         }
         .padding(.horizontal, profileHorizontalPadding)
         .zIndex(1)
@@ -106,13 +113,12 @@ struct SEMyProfileView: View {
               GridItem(.flexible()),
               GridItem(.flexible()),
             ], spacing: 20) {
-              ForEach(Array(apiViewModel.myFeed.enumerated()), id: \.element) { index, content in
+              ForEach(Array(apiViewModel.myFeed.enumerated()), id: \.element) { index , content in
                 NavigationLink {
-                  MyFeedView(currentIndex: index)
-
+                  MyFeedKitView(index: index)
                 } label: {
                   videoThumbnailView(
-                    thumbnailUrl: content.thumbnailUrl ?? "",
+                    thumbnailURL: content.thumbnailUrl ?? "",
                     viewCount: content.contentViewCount ?? 0)
                 }
               }
@@ -128,7 +134,7 @@ struct SEMyProfileView: View {
           .coordinateSpace(name: "SCROLL")
           .zIndex(0)
           Spacer()
-        // 북마크 탭 & 올린 컨텐츠 있음
+        // O 탭 & 올린 컨텐츠 있음
         case (.bookmark, _, false):
           ScrollView {
             LazyVGrid(columns: [
@@ -140,7 +146,7 @@ struct SEMyProfileView: View {
                 NavigationLink {
                   BookmarkedFeedView(currentIndex: index)
                 } label: {
-                  videoThumbnailView(thumbnailUrl: content.thumbnailUrl, viewCount: content.viewCount)
+                  videoThumbnailView(thumbnailURL: content.thumbnailUrl, viewCount: content.viewCount)
                 }
               }
             }
@@ -188,13 +194,11 @@ struct SEMyProfileView: View {
         Divider().background(Color("Gray10"))
         NavigationLink {
           NotificationSettingView()
-
         } label: {
           bottomSheetRowWithIcon(systemName: "bell", text: "알림")
         }
         NavigationLink {
           LegalInfoView()
-
         } label: {
           bottomSheetRowWithIcon(systemName: "info.circle", text: "약관 및 정책")
         }
@@ -211,11 +215,8 @@ struct SEMyProfileView: View {
         }
         NavigationLink {
           GuideStatusView()
-
         } label: {
-          bottomSheetRowWithIcon(
-            systemName: "exclamationmark.triangle.fill",
-            text: "신고")
+          bottomSheetRowWithIcon(systemName: "exclamationmark.triangle.fill", text: "가이드 상태")
         }
         Group {
           Divider().background(Color("Gray10"))
@@ -223,7 +224,7 @@ struct SEMyProfileView: View {
             withAnimation {
               bottomSheetPosition = .hidden
             }
-            alertViewModel.linearAlert(title: "정말 로그아웃하시겠어요?", destructiveText: "로그아웃") {
+            alertViewModel.linearAlert(title: "정말 로그아웃하시겠어요?", cancelText: "취소", destructiveText: "로그아웃") {
               apiViewModel.reset()
               GIDSignIn.sharedInstance.signOut()
               userAuth.appleSignout()
@@ -239,6 +240,7 @@ struct SEMyProfileView: View {
             alertViewModel.linearAlert(
               title: "정말 삭제하시겠어요?",
               content: "삭제하시면 회원님의 모든 정보와 활동 기록이 삭제됩니다. 삭제된 정보는 복구할 수 없으니 신중하게 결정해주세요.",
+              cancelText: "취소",
               destructiveText: "삭제")
             {
               Task {
@@ -277,35 +279,33 @@ struct SEMyProfileView: View {
   }
 }
 
-extension SEMyProfileView {
+extension MyProfileView {
   @ViewBuilder
   func profileInfo() -> some View {
     VStack(spacing: 0) {
-      Spacer().frame(height: 48)
+      Spacer().frame(height: 64)
       profileImageView(url: apiViewModel.myProfile.profileImage, size: profileImageSize)
-        .padding(.bottom, 12)
+        .padding(.bottom, 16)
       Text(apiViewModel.myProfile.userName)
-        .font(.system(size: 18, weight: .semibold).width(.expanded))
         .foregroundColor(Color.LabelColor_Primary_Dark)
-        .frame(height: 28)
-      Spacer().frame(minHeight: 10)
+        .fontSystem(fontDesignSystem: .title2_Expanded)
+        .padding(.bottom, 4)
+      Spacer()
       Color.clear.overlay {
         Text(apiViewModel.myProfile.introduce ?? "")
           .foregroundColor(Color.LabelColor_Secondary_Dark)
-          .font(.system(size: 14, weight: .regular))
           .fontSystem(fontDesignSystem: .body2_KO)
           .lineLimit(nil)
           .multilineTextAlignment(.center)
           .fixedSize(horizontal: false, vertical: true)
           .scaleEffect(introduceScale)
       }
-      .frame(height: introduceHeight) // 20 max
-      .padding(.bottom, 8)
+      .frame(height: introduceHeight)
+      .padding(.bottom, 16)
       .padding(.horizontal, 48)
       Spacer()
       NavigationLink {
         ProfileEditView()
-
       } label: {
         Text("프로필 편집")
           .fontSystem(fontDesignSystem: .subtitle2_KO)
@@ -314,41 +314,41 @@ extension SEMyProfileView {
           .frame(width: profileEditButtonWidth, height: profileEditButtonHeight)
       }
       .frame(width: profileEditButtonWidth, height: profileEditButtonHeight)
-      .padding(.bottom, 16)
+      .padding(.bottom, 24)
       .buttonStyle(ProfileEditButtonStyle())
-      HStack(spacing: 48) {
+      HStack(spacing: 0) {
         VStack(spacing: 4) {
           Text("\(apiViewModel.myWhistleCount)")
             .foregroundColor(Color.LabelColor_Primary_Dark)
-            .font(.system(size: 16, weight: .semibold).width(.expanded))
+            .fontSystem(fontDesignSystem: .title2_Expanded)
             .scaleEffect(whistleFollowerTextScale)
-          Text("whistle")
+          Text("휘슬")
             .foregroundColor(Color.LabelColor_Secondary_Dark)
-            .font(.system(size: 10, weight: .semibold))
+            .fontSystem(fontDesignSystem: .caption_SemiBold)
             .scaleEffect(whistleFollowerTextScale)
         }
-        Rectangle().frame(width: 1, height: .infinity).foregroundColor(.white)
+        .hCenter()
+        Rectangle().frame(width: 1).foregroundColor(.white).scaleEffect(0.5)
         NavigationLink {
           MyFollowListView()
-
         } label: {
           VStack(spacing: 4) {
-            Text("\(apiViewModel.myFollow.followerCount)")
+            Text("\(filteredFollower.count)")
               .foregroundColor(Color.LabelColor_Primary_Dark)
-              .font(.system(size: 16, weight: .semibold).width(.expanded))
+              .fontSystem(fontDesignSystem: .title2_Expanded)
               .scaleEffect(whistleFollowerTextScale)
-            Text("follower")
+            Text("팔로워")
               .foregroundColor(Color.LabelColor_Secondary_Dark)
-              .font(.system(size: 10, weight: .semibold))
+              .fontSystem(fontDesignSystem: .caption_SemiBold)
               .scaleEffect(whistleFollowerTextScale)
           }
+          .hCenter()
         }
       }
-      .frame(height: whistleFollowerTabHeight) // 42 max
-      .padding(.bottom, 10)
-      Spacer()
+      .frame(height: whistleFollowerTabHeight)
+      .padding(.bottom, 32)
     }
-    .frame(height: 278 + (146 * progress))
+    .frame(height: 418 + (240 * progress))
     .frame(maxWidth: .infinity)
     .overlay {
       VStack(spacing: 0) {
@@ -361,7 +361,7 @@ extension SEMyProfileView {
           } label: {
             Circle()
               .foregroundColor(.Gray_Default)
-              .frame(width: 40, height: 40)
+              .frame(width: 48, height: 48)
               .overlay {
                 Image(systemName: "ellipsis")
                   .resizable()
@@ -371,21 +371,19 @@ extension SEMyProfileView {
                   .frame(width: 20, height: 20)
               }
           }
-          .offset(y: 28 - topSpacerHeight)
-          .padding(.top, 16)
+          .offset(y: 64 - topSpacerHeight)
           .padding(.horizontal, 16 - profileHorizontalPadding)
         }
         Spacer()
       }
-      .padding(.horizontal, 16)
-      .padding(.top, 8)
+      .padding(16)
     }
   }
 
   @ViewBuilder
-  func videoThumbnailView(thumbnailUrl: String, viewCount: Int) -> some View {
+  func videoThumbnailView(thumbnailURL: String, viewCount: Int) -> some View {
     Color.black.overlay {
-      KFImage.url(URL(string: thumbnailUrl))
+      KFImage.url(URL(string: thumbnailURL))
         .placeholder {
           Color.black
         }
@@ -408,7 +406,7 @@ extension SEMyProfileView {
         .frame(maxWidth: .infinity, alignment: .leading)
       }
     }
-    .frame(height: UIScreen.getHeight(204))
+    .frame(height: 204)
     .cornerRadius(12)
   }
 
@@ -433,9 +431,19 @@ extension SEMyProfileView {
     .padding(.bottom, 76)
     Spacer()
   }
+
+  @ViewBuilder
+  func bookmarkEmptyView() -> some View {
+    Spacer()
+    Text("저장한 콘텐츠가 없습니다")
+      .fontSystem(fontDesignSystem: .body1_KO)
+      .foregroundColor(.LabelColor_Primary_Dark)
+      .padding(.bottom, 64)
+    Spacer()
+  }
 }
 
-extension SEMyProfileView {
+extension MyProfileView {
   @ViewBuilder
   func bottomSheetRowWithIcon(
     systemName: String,
@@ -472,23 +480,13 @@ extension SEMyProfileView {
     .frame(height: 56)
     .padding(.horizontal, 16)
   }
-
-  @ViewBuilder
-  func bookmarkEmptyView() -> some View {
-    Spacer()
-    Text("저장한 콘텐츠가 없습니다")
-      .fontSystem(fontDesignSystem: .body1_KO)
-      .foregroundColor(.LabelColor_Primary_Dark)
-      .padding(.bottom, 64)
-    Spacer()
-  }
 }
 
 // MARK: - Sticky Header Computed Properties
 
-extension SEMyProfileView {
+extension MyProfileView {
   var progress: CGFloat {
-    -(offsetY / 132) > 1 ? -1 : (offsetY > 0 ? 0 : (offsetY / 132))
+    -(offsetY / 177) > 1 ? -1 : (offsetY > 0 ? 0 : (offsetY / 177))
   }
 
   var progressOpacity: CGFloat {
@@ -499,8 +497,8 @@ extension SEMyProfileView {
     switch -offsetY {
     case ..<0:
       16
-    case 0 ..< 28:
-      16 + (16 * (offsetY / 28))
+    case 0 ..< 64:
+      16 + (16 * (offsetY / 64))
     default:
       0
     }
@@ -510,8 +508,8 @@ extension SEMyProfileView {
     switch -offsetY {
     case ..<0:
       32
-    case 0 ..< 28:
-      32 + (32 * (offsetY / 28))
+    case 0 ..< 64:
+      32 + (32 * (offsetY / 64))
     default:
       0
     }
@@ -520,9 +518,9 @@ extension SEMyProfileView {
   var topSpacerHeight: CGFloat {
     switch -offsetY {
     case ..<0:
-      28
-    case 0 ..< 28:
-      28 + offsetY
+      64
+    case 0 ..< 64:
+      64 + offsetY
     default:
       0
     }
@@ -531,9 +529,9 @@ extension SEMyProfileView {
   var profileImageSize: CGFloat {
     switch -offsetY {
     case ..<0:
-      56
-    case 0 ..< 68:
-      56 + (56 * (offsetY / 68))
+      100
+    case 0 ..< 122:
+      100 + (100 * (offsetY / 122))
     default:
       0
     }
@@ -541,10 +539,10 @@ extension SEMyProfileView {
 
   var whistleFollowerTabHeight: CGFloat {
     switch -offsetY {
-    case ..<68:
-      42
-    case 68 ..< 126:
-      42 + (42 * ((offsetY + 68) / 58))
+    case ..<122:
+      54
+    case 122 ..< 200:
+      54 + (54 * ((offsetY + 122) / 78))
     default:
       0
     }
@@ -554,8 +552,8 @@ extension SEMyProfileView {
     switch -offsetY {
     case ..<122:
       1
-    case 68 ..< 126:
-      1 - abs((offsetY + 68) / 58)
+    case 122 ..< 200:
+      1 - abs((offsetY + 122) / 78)
     default:
       0
     }
@@ -563,10 +561,10 @@ extension SEMyProfileView {
 
   var profileEditButtonHeight: CGFloat {
     switch -offsetY {
-    case ..<126:
+    case ..<200:
       36
-    case 126 ..< 146:
-      28 + (28 * ((offsetY + 126) / 20))
+    case 200 ..< 252:
+      36 + (36 * ((offsetY + 200) / 52))
     default:
       0
     }
@@ -574,10 +572,10 @@ extension SEMyProfileView {
 
   var profileEditButtonWidth: CGFloat {
     switch -offsetY {
-    case ..<126:
+    case ..<200:
       114
-    case 126 ..< 146:
-      79 + (79 * ((offsetY + 126) / 20))
+    case 200 ..< 252:
+      114 + (114 * ((offsetY + 200) / 52))
     default:
       0
     }
@@ -585,10 +583,10 @@ extension SEMyProfileView {
 
   var profileEditButtonScale: CGFloat {
     switch -offsetY {
-    case ..<126:
+    case ..<200:
       1
-    case 126 ..< 146:
-      1 - abs((offsetY + 126) / 20)
+    case 200 ..< 252:
+      1 - abs((offsetY + 200) / 52)
     default:
       0
     }
@@ -596,10 +594,10 @@ extension SEMyProfileView {
 
   var introduceHeight: CGFloat {
     switch -offsetY {
-    case ..<146:
+    case ..<252:
       20
-    case 146 ..< 202:
-      20 + (20 * ((offsetY + 146) / 56))
+    case 252 ..< 305:
+      20 + (20 * ((offsetY + 252) / 53))
     default:
       0
     }
@@ -607,10 +605,10 @@ extension SEMyProfileView {
 
   var introduceScale: CGFloat {
     switch -offsetY {
-    case ..<146:
+    case ..<252:
       1
-    case 146 ..< 202:
-      1 - abs((offsetY + 146) / 56)
+    case 252 ..< 305:
+      1 - abs((offsetY + 252) / 53)
     default:
       0
     }
@@ -618,12 +616,12 @@ extension SEMyProfileView {
 
   var tabOffset: CGFloat {
     switch -offsetY {
-    case ..<146:
+    case ..<252:
       0
-    case 146 ..< 202:
-      32 * ((offsetY + 146) / 56)
-    case 202...:
-      -32
+    case 252 ..< 305:
+      36 * ((offsetY + 252) / 53)
+    case 305...:
+      -36
     default:
       0
     }
@@ -631,11 +629,11 @@ extension SEMyProfileView {
 
   var tabPadding: CGFloat {
     switch -offsetY {
-    case ..<146:
+    case ..<252:
       16
-    case 146 ..< 202:
-      8 + (8 * ((offsetY + 146) / 56))
-    case 202...:
+    case 252 ..< 305:
+      16 + (16 * ((offsetY + 252) / 53))
+    case 305...:
       0
     default:
       0
@@ -644,11 +642,11 @@ extension SEMyProfileView {
 
   var tabHeight: CGFloat {
     switch -offsetY {
-    case ..<146:
+    case ..<252:
       48
-    case 146 ..< 202:
-      48 + (48 * ((offsetY + 146) / 56))
-    case 202...:
+    case 252 ..< 305:
+      48 + (48 * ((offsetY + 252) / 53))
+    case 305...:
       0
     default:
       0
@@ -656,6 +654,12 @@ extension SEMyProfileView {
   }
 
   var videoOffset: CGFloat {
-    offsetY < -202 ? 202 : -offsetY
+    offsetY < -305 ? 305 : -offsetY
+  }
+}
+
+extension MyProfileView {
+  var filteredFollower: [FollowerData] {
+    apiViewModel.myFollow.followerList.filter { !BlockList.shared.userIds.contains($0.followerId) }
   }
 }
